+++ src/nuxt/index.ts
@@ -1,0 +1,2 @@
+export { default } from './module'
+export type { IOrigamNuxtModuleOptions, IOrigamNuxtRuntimeConfig } from '../interfaces'